requirements.txt
fastapi
uvicorn[standard]
Jinja2
python-multipart
numpy
Pillow
tifffile
imageio
pillow-heif
zipstream-ng
opencv-python